+++ pkg/backup/restore.go
@@ -10,6 +10,7 @@ import (
 	"os"
 	"os/exec"
 	"path"
+	"regexp"
 	"strings"
 	"time"
 
@@ -251,6 +252,8 @@ func RestoreSchema(cfg *config.Config, ch *clickhouse.ClickHouse, backupName str
 	return nil
 }
 
+var UUIDWithReplicatedMergeTreeRE = regexp.MustCompile(`^(.+)(UUID)(\s+)'([^']+)'(.+)({uuid})(.*)`)
+
 func createTables(cfg *config.Config, ch *clickhouse.ClickHouse, tablesForRestore ListOfTables, version int, log *apexLog.Entry) error {
 	totalRetries := len(tablesForRestore)
 	restoreRetries := 0
@@ -269,6 +272,14 @@ func createTables(cfg *config.Config, ch *clickhouse.ClickHouse, tablesForRestor
 			schema.Query = strings.Replace(
 				schema.Query, "CREATE WINDOW VIEW", "ATTACH WINDOW VIEW", 1,
 			)
+			// https://github.com/AlexAkulov/clickhouse-backup/issues/466
+			if cfg.General.RestoreSchemaOnCluster == "" && strings.Contains(schema.Query, "{uuid}") && strings.Contains(schema.Query, "Replicated") {
+				if !strings.Contains(schema.Query, "UUID") {
+					log.Warnf("table query doesn't contains UUID, can't guarantee properly restore for ReplicatedMergeTree")
+				} else {
+					schema.Query = UUIDWithReplicatedMergeTreeRE.ReplaceAllString(schema.Query, "$1$2$3'$4'$5$4$7")
+				}
+			}
 			restoreErr = ch.CreateTable(clickhouse.Table{
 				Database: schema.Database,
 				Name:     schema.Table,